Split autoModeration types
